Shoot ball

+++ BrickBreak/src/Ball.h
@@ -10,6 +10,6 @@ public:
     Ball();
     Ball(const glm::vec2& pos, float radius, glm::vec2 velocity, Texture* texture);
 
-    glm::vec2 move(float delta, unsigned int windowWidth);
+    glm::vec2 move(double delta, unsigned int windowWidth);
     void reset(const glm::vec2& pos, const glm::vec2& velocity);
 };
